Update on updateSwapsTransaction fn
